Refactor management into sub package
management/service_manager.py
management/endpoint.py
management/reserved_ip.py
management/request_result.py
management/data_disk.py
